ensure that histogram data is cast as floats

+++ library/ml_defs.py
@@ -265,4 +265,4 @@ def cum_poisson_nll(y_true,y_pred):
     negative log-likelihood loss function for
     Poisson observations
     """
-    return K.sum(K.cumsum(y_pred,axis=1)-K.cumsum(y_true,axis=1)*K.log(K.cumsum(y_pred,axis=1)+1e-9),axis=1)
+    return K.sum(K.cumsum(y_pred,axis=1)-K.cumsum(K.cast(y_true,'float32'),axis=1)*K.log(K.cumsum(y_pred,axis=1)+1e-9),axis=1)
